pyproject.toml
[tool.poetry]
name = "opponent-adjusted-metrics"
version = "0.1.0"
description = "End-to-end contextual, opponent-adjusted football metrics using StatsBomb Open Data"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "opponent_adjusted", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.0"
numpy = "^1.26.0"
pyarrow = "^15.0.0"
scipy = "^1.12.0"
lightgbm = "^4.3.0"
scikit-learn = "^1.4.0"
joblib = "^1.3.2"
sqlalchemy = "^2.0.25"
psycopg = {extras = ["binary"], version = "^3.1.18"}
alembic = "^1.13.1"
pydantic = "^2.6.0"
matplotlib = "^3.8.2"
seaborn = "^0.13.2"
plotly = "^5.18.0"
uvicorn = "^0.27.0"
fastapi = "^0.109.0"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.12.0"
mplsoccer = "^1.6.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
black = "^24.1.0"
ruff = "^0.2.0"
mypy = "^1.8.0"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
